Refactored.
Removed static version of GetDbType() method and moved the body
to the overridden version. Since the dataScale and dataPrecision
params were no longer being used, there was no longer any reason
for the static version to exist. Also, it looks like the overridden
version was still being referenced and it did not contain the correct
mappings.